增加发布静态文件的脚本

+++ os/linux/prod-app-dir/app/scp.sh
@@ -3,14 +3,14 @@
 APP=qh-wap
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 
-dataStr=$(date +%Y%m%d%H%M%S)
-
 if [ -e ${DIR}/upload/*.war ] ; then
     for f in ${DIR}/upload/*.war
     do
         md5Str=$(md5sum ${f} | cut -f1 -d ' ')
-        bakName="$(basename $f).${dataStr}.$md5Str"
-        mv $f $DIR/bak/$bakName
+        bakName="$(basename $f).$md5Str"
+        if [ ! -e $DIR/bak/$bakName ] ; the
+            mv $f $DIR/bak/$bakName
+        fi
     done
 fi
